Fri, Sep 27, 2024, 10:04 AM -04:00

+++ readability/readability.c
@@ -32,7 +32,7 @@ int main (void)
         float number_of_characters = count_alpha(text);
         float spaces_plus_one = no_of_spaces++;
 
-        L = L + ( number_of_characters / spaces_plus_one );
+        L = L + ( (float)number_of_characters / spaces_plus_one );
 
         printf("%f\n", L);
